+++ api/app/blueprints/interest/routes.py
@@ -55,6 +55,8 @@ def update_interest(_id):
 def delete_interest(_id):
     try:
         interest = Interest.query.get(_id)
+        if interest is None:
+            return make_response("Interest with id={id} not found".format(id=_id), 404)
         db.session.delete(interest)
         db.session.commit()
     except exc.SQLAlchemyError as exception_message:
